requirements.txt
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
bleak==0.21.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.4
dbus-fast==2.12.0
exceptiongroup==1.1.3
fastapi==0.103.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
protobuf==4.24.4
pycparser==2.21
pydantic==2.4.2
pydantic-core==2.10.1
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2
netifaces==0.11.0